Move particle component to separate file
